% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{ttcg-package}
\alias{ttcg-package}
\alias{_PACKAGE}
\title{ttcg: Three-term conjugate gradient minimization algorithms}
\description{
Some accelerated three-term conjugate gradient algorithms implemented purely in R with the same user interface as optim(). The search directions and acceleration scheme are described in Andrei, N. (2013) <doi:10.1016/j.amc.2012.11.097>, Andrei, N. (2013) <doi:10.1016/j.cam.2012.10.002>, and Andrei, N (2015) <doi:10.1007/s11075-014-9845-9>. Line search is done by a hybrid algorithm incorporating the ideas in Oliveia and Takahashi (2020) <doi:10.1145/3423597> and More and Thuente (1994) <doi:10.1145/192115.192132>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.sr.ht/~hckiang/ttcg}
}

}
\author{
Hao Chi Kiang, \email{hello@hckiang.com}
}
