% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{ttcg}
\alias{ttcg}
\title{ttcg: Three-term conjugate gradient minimization algorithms}
\usage{
ttcg(par, fn, gr = NULL, method = "TTDES", control = list(), ...)
}
\arguments{
\item{par}{A numerical vector containing the initial parameters.}

\item{fn}{A function to be minimized. It should accept a numerical vector as its sole
argument and return a scaler.}

\item{gr}{The gradient function of \code{fn}. It should accept the same argument as \code{fn}
and return a vector of the same length as \code{par}.}

\item{method}{A character string, one of 'TTDES', 'TTCG', 'THREECG'. This determines how the
line search direction is computed. 'TTDES' is the default method.}

\item{control}{A list of control parameters. See Details.}

\item{...}{Extra arguments to be passed to \code{fn}}
}
\value{
A list containing the following named elements.
              \describe{
                \item{par}{The optimal parameter.}
                \item{value}{The optimal function value.}
                \item{counts}{An integer vector containing the number of function and gradient calls
                              used during the optimization.}
                \item{convergence}{An integer indicating convergence status. '0' means successful convergence; '1'
                                   means \code{maxit} has been reached; '2' means a line search failure in which
                                   a point that satisfies the weak Wolfe condition is not found (perhaps the function
                                   is unbounded below).}
                \item{message}{A character string giving additional message.}
              }
}
\description{
The ttcg package implements several Neculai Andrei's three-term conjugate gradient algorithms.

The \code{ttcg} function minimizes a given function using several Neculai Andrei's
three-term conjugate gradient algorithms. Line search is done by a bisection-like
weak-Wolfe search inspired by Oliveira and Takahashi's interpolate-truncate-project
algorithm and More-Thuente algorithm.
}
\details{
The \code{control} argument is a list that can contain any of the following named
element:

    \describe{
      \item{maxit}{The maximal number of iteration.}
      \item{gl2tol}{A positive small number. The iteration will be terminated if the
                    squared Euclidean norm is smaller than this number. Default is
                    \code{min(1e-5, length(par))*1e-7)}. To turn off this test,
                    set it to any negative values.}
      \item{gmaxtol}{A positive small number. The iteration will be terminated if the infinity norm of the graident is smaller than
                     \code{gmaxtol}. Default is 1e-6. To turn off this test, set it to any negative values.}
      \item{ftol}{A positive small number. The iteration will be terminated if \eqn{|f_{k+1} - f_k| < ftol * (1 + |f_k|)}.
                  To turn off this test, set it to any negative values.}
      \item{c1}{The line search parameter for the sufficient descent condition. Default is 1e-4.}
      \item{c2}{The line search parameter for the curvature condition. Default is 0.08.}
      \item{trace}{Either TRUE or FALSE, indicating whether or not details should be printed to the terminal
                   during the optimization. Default is FALSE.}
    }
}
\references{
Andrei, N. (2013). On three-term conjugate gradient algorithms for unconstrained optimization. Applied Mathematics and Computation, 219(11), 6316-6327.

Andrei, N. (2013). A simple three-term conjugate gradient algorithm for unconstrained optimization. Journal of Computational and Applied Mathematics, 241, 19-29.

Andrei, N. (2015). A new three-term conjugate gradient algorithm for unconstrained optimization. Numerical Algorithms, 68(2), 305-321.

Oliveira, I. F., & Takahashi, R. H. (2020). An Enhancement of the Bisection Method Average Performance Preserving Minmax Optimality. ACM Transactions on Mathematical Software (TOMS), 47(1), 1-24.

More, J. J., & Thuente, D. J. (1994). Line search algorithms with guaranteed sufficient decrease. ACM Transactions on Mathematical Software (TOMS), 20(3), 286-307.
}
\author{
Hao Chi Kiang, \email{hello@hckiang.com}
}
